156/delete page (#176)
* WIP
* animate delete page button on hover/focus

* styling improvements to nav

* increase padding to accomodate delete button
* lightened delete button text color
* fix truncated text for long page titles
* ensure PageItem bg is shown on hover of delete button

* redirect to correct page on delete

if there are 3 pages. if you delete the second you should go to page 1, if you delete the first you should go to page 2

* trying to get new page creation working

* handle deletion of last page in section

* add tests for deletion

* fix linting issues, improve a test

* fix deletePage in mock API

* create new page when creating new section

* add aria-label to delete buttons

* rename CloseButton -> DeleteButton

* fix tests

* _really_ fix tests

* add autoFocus to QuestionnaireMeta

* fade page nav item before delete

* distinguishes deletion from adding new page

* improve test for deleting pages

* avoid re-rendering QuestionnaireDesignPage if page is null/undefined

Page is null/undefined when the current page has just been deleted, and the redirect yet to happen. This avoids having to check for null further down the tree, and allows us to have all the regular animations for deletions