Write a test to catch any regression of issue #191.
There was an issue in a previous version of the application where
multiple choice answer options were being lost as soon as a new
option was being added.

Significant refactoring has happened and the issue has not been
seen since.

This test has been added retrospectively to ensure that we do
not introduce any regression and re-introduce the problem.

Unfortunately, due to some unknown issue, the test does not
run on Travis using Chrome Headless. The PR can be safely
merged, since the test will be skipped until the issue with
running the test on Travis is better understood.