Add an accordion component.
Introduce new accordion component into the application which displays
the page and ansswer properties.

At this stage the properties will not be persisted, this will be done
in a future task.

If the accordions fill the panel such as a scroll bar is displayed, then
the panel should slide into view when opened.