Tech Debt - Stop rebuilding on docker start
This implements having a way for the app to take docker environment
variables and use them without having to rebuild them each time.

It does this by rewriting just the index.html on docker start and
the using the values defined there as the configuration options
throughout the rest of the app.