Bug - Full story not fully enabled
Full story was partially using build time and runtime environment
variables.

Build time for the script include and runtime to actually hook our
identity to it.

This resulted in an error when the build time was set to false and
runtime was set to true.

To resolve this we now only use runtime environment variables to
determine whether to use full story.