Update withEntityEditor to handle slow network requests
During the piping cypress spec whilst creating metadata in the piping
spec some of the text would be cut off from the metadata alias. This was
because the network request for updating the metadata key had come back
whilst cypress was updating the alias - this caused the entity managed
by withEntityEditor to be updated and hence the field was cleared out.

The solution to this was to track the field that was dirty and to
"merge" it into the returned data so the user's changes aren't lost.